% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date_details.R
\name{add_date_details}
\alias{add_date_details}
\title{Extract observation date details}
\usage{
add_date_details(df)
}
\arguments{
\item{df}{Input data frame from water tracker with added field "obsdate" from
\code{\link[=estimate_obsdate]{estimate_obsdate()}}}
}
\value{
tibble with added fields "year", "month", "wateryear", and
"month_name".
}
\description{
Extract observation date details
}
\details{
Internal to \code{\link[=format_watertracker]{format_watertracker()}}. Not meant to be called
separately.
}
