% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_watertracker.R
\name{analyze_watertracker}
\alias{analyze_watertracker}
\title{Analyze Water Tracker data}
\usage{
analyze_watertracker(
  df,
  prob = 0.95,
  delta = 0.2,
  full_prop_threshold = 0.55,
  wet_prop_threshold = 0.2,
  fullmode = "M"
)
}
\arguments{
\item{df}{tibble resulting from \code{\link[=format_watertracker]{format_watertracker()}}}

\item{prob}{Single numberic value passed to \code{\link[=quantile]{quantile()}}; see Details in
\code{\link[=estimate_floodstatus]{estimate_floodstatus()}}.}

\item{delta}{Numeric proportion change representing significant increase or decrease in flooding.}

\item{full_prop_threshold}{Numeric value (0-1), see Details in
\code{\link[=estimate_floodstatus]{estimate_floodstatus()}}.}

\item{wet_prop_threshold}{Numeric value (0-1), see Details in
\code{\link[=estimate_floodstatus]{estimate_floodstatus()}}.}

\item{fullmode}{One of \code{M} or \code{H}; default operational mode for units
considered fully flooded; see \code{\link[=estimate_wetlandmode]{estimate_wetlandmode()}} for details.}
}
\value{
tibble with fields: \code{unit}, \code{wateryear}, \code{month}, \code{month_name},
\code{year}, \code{PercentWater} (the maximum observed in the given month and year),
\code{PercentWater_max} (the maximum observed over all dates in the dataset),
\code{flood_prop} (\code{PercentWater}/\code{PercentWater_max}), and \code{flood_status}.
}
\description{
Interprets sequence of Water Tracker data in each unit to interpret flooding
status and assign monthly water management modes for use with the \verb{Wetland Water Budget Tool} a simple spread-sheet based tool allowing non-specialists
to estimate wetland water budgets from a combination of management unit size
and schedule of monthly operational modes.
}
\details{
Wrapper function that calls \code{\link[=estimate_floodstatus]{estimate_floodstatus()}} and
\code{\link[=estimate_wetlandmode]{estimate_wetlandmode()}}. See details of each of those functions for more
information.
}
\examples{
#format_watertracker(sampledat) |> analyze_watertracker()

}
