% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_area.R
\name{estimate_area}
\alias{estimate_area}
\title{Estimate area}
\usage{
estimate_area(df)
}
\arguments{
\item{df}{tibble resulting from \code{\link[=format_watertracker]{format_watertracker()}}}
}
\value{
tibble with additional field "area_ac"
}
\description{
Estimate area
}
\examples{
#estimate_area(df)
}
