% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_floodstatus.R
\name{estimate_floodstatus}
\alias{estimate_floodstatus}
\title{Estimate flooding status}
\usage{
estimate_floodstatus(df, prob, delta, full_prop_threshold, wet_prop_threshold)
}
\arguments{
\item{df}{Input tibble from \code{\link[=format_watertracker]{format_watertracker()}}.}

\item{prob}{Numeric value passed to \code{\link[=quantile]{quantile()}}; see Details.}

\item{delta}{Numeric proportion change representing significant increase or
decrease in flooding.}

\item{full_prop_threshold}{Numeric value (0-1); see Details.}

\item{wet_prop_threshold}{Numeric value (0-1); see Details.}
}
\value{
tibble with added fields: \code{ObservedAreaWater_pq},
\code{ObservedAreaWater_adjust}, \code{flood_delta}, \code{flood_trend}, \code{flood_prop}, and
\code{flood_status}
}
\description{
Internal to \code{\link[=analyze_watertracker]{analyze_watertracker()}}. Not meant to be called separately.
}
\details{
For each unit, this function first calculates \code{ObservedAreaWater_pq}
as a percentile value of 'ObservedAreaWaterHa' (provided in the raw Water
Tracker data) over all observation dates in the dataset, as determined by
\code{prob}. This value is intended to represents the typical upper limit of
wetted area, to account for upland areas in the unit that are not typically
intentionally flooded as part of regular manaagement. The value for \code{prob}
should typically be set less than 1 to allow for outliers with extreme
flooding events. Note that providing multiple values of \code{prob}, while
supported by the \code{\link[=quantile]{quantile()}} function, is not supported by this function.

The \code{ObservedAreaWater_adjust} is then calculated to reflect this upper
limit of wetted area, such that values of \code{ObservedAreaWaterHa} that exceed
the are reduced to this value. Likewise, \code{flood_prop} is calculated as the
proportion of \code{ObservedAreaWater_adjust} relative to this upper limit.

The proportion change in flooded area from the previous observation date is
then calculated (\code{flood_delta}) and categorized (\code{flood_trend}) by
direction and magnitude of change, based on the input value for \code{delta}
provided, as:
\itemize{
\item \code{+}:  \code{flood_delta} >= \code{delta}
\item \verb{(+)}: \code{delta} > \code{flood_delta} > 0
\item \code{0}: \code{flood_delta} = 0
\item \verb{(-)}: 0 > \code{flood_delta} > -\code{delta}
\item \code{-}: -\code{delta} >= \code{flood_delta}
}

Finally the \code{flood_status} on each observation date is then designated as:
\itemize{
\item \code{full}: \code{flood_prop} >= \code{full_prop_threshold}
\item \code{wet}: \code{wet_prop_threshold} <= \code{flood_prop} < \code{full_prop_threshold},
\item \code{trace}: \code{flood_prop} > 0 but < \verb{wet prop threshold}
\item \code{dry}: \code{flood_prop} = 0
}
}
