% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_modes.R
\name{interpret_modes}
\alias{interpret_modes}
\title{Interpret wetland operational modes}
\usage{
interpret_modes(df, wetmode = "M", wetthreshold = 50, obsthreshold = 80)
}
\arguments{
\item{df}{tibble resulting from \code{\link[=estimate_area]{estimate_area()}}}

\item{wetmode}{Default operational mode for units determined to be flooded}

\item{wetthreshold}{Threshold applied to the \code{PercentWater} field for
considering a wetland unit to be flooded}

\item{obsthreshold}{Threshold applied to the \code{PercentObserved} field for
considering data to be useful and included}
}
\value{
tibble with additional field "mode"
}
\description{
Interpret flooding status on each date and monthly operational modes from
Water Tracker data
}
\details{
This function evaluates the Water Tracker data to first determine
whether a wetland management unit should be considered flooded or not on a
given observation date, and then interpret the flooding status on one or
more observation dates within each month to estimate the monthly
operational mode.

Flooding status is evaluated by setting the \code{wetthreshold} to a value above
which the unit will be considered flooded (\code{status = "wet"}) and below
which the unit will be considered dry (\code{status = "dry"}). However, if
PercentObserved is below the \code{obsthreshold}, the data will be considered
unreliable and the status changed to \code{NA}.

Operational mode codes are derived from the \verb{Wetland Water Budget Tool} a
simple spread-sheet based tool allowing non-specialists to estimate wetland
water budgets from a combination of management unit size and schedule of
monthly operational modes.

Possible values include:
\itemize{
\item N: No water supply
\item F: Flood-up
\item D: Drawdown
\item M: Maintenance (actively maintained at full)
\item H: Hold (filled but not actively maintained)
\item I: Irrigation
\item C: Cropped (no water supply, but actively growing vegetation)
}

Monthly operational modes are evaluated by examining the flooding status
for all dates within a single month for each management unit ("unit" by
default) and water year. If all are \code{"dry"}, \code{mode = "N"} (no water
supply)). If all are \code{"wet"}, the mode is set equal to the \code{wetmode}, which
should be either "M" or "H". If the first date is \code{"dry"} and the last date
is \code{"wet"}, \code{mode = "F"}, while for the reverse, \verb{"mode = "D"}. Any other
combinations will result in \code{mode = "UNKNOWN"} and should be manually
inspected.
}
\examples{
format_watertracker(sampledat) |> estimate_area() |> interpret_modes()
}
