% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_area.R
\name{estimate_area}
\alias{estimate_area}
\title{Estimate area}
\usage{
estimate_area(df)
}
\arguments{
\item{df}{tibble resulting from \code{\link[=format_watertracker]{format_watertracker()}}}
}
\value{
tibble with additional field "area_ac"
}
\description{
Estimate the area of each wetland management unit from the area observed in
the Water Tracker data.
}
\details{
This function converts the "ObservedArea" field (m^2) in the
original Water Tracker output data to an estimate of the total area of each
management unit, in acres. The area is estimated as the maximum of all
ObservedArea values with the unit name.
}
\examples{
format_watertracker(sampledat) |> estimate_area()
}
