% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_watertracker.R
\name{format_watertracker}
\alias{format_watertracker}
\title{Format Water Tracker data}
\usage{
format_watertracker(df, unitID = "Name")
}
\arguments{
\item{df}{Input data frame from water tracker}

\item{unitID}{Character string for the field containing the management units
for which hydrological data will be evaluated}
}
\value{
tibble with added fields unit, year, month, wateryear, month_name,
midpoint
}
\description{
Format Water Tracker data
}
\examples{
# format_watertracker(df)

}
