% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_watertracker.R
\name{format_watertracker}
\alias{format_watertracker}
\title{Format Water Tracker data}
\usage{
format_watertracker(df, unitID = "Name", wetthreshold = 50, obsthreshold = 80)
}
\arguments{
\item{df}{Input data frame from water tracker}

\item{unitID}{Character string for the field containing the management units
for which hydrological data will be evaluated}

\item{wetthreshold}{Threshold applied to proportion open water for
considering a wetland unit to be flooded}

\item{obsthreshold}{Threshold applied to the proportion observed for
including water tracker data}
}
\value{
tibble with added fields year, month, wateryear, month_name, status,
midpoint
}
\description{
Format Water Tracker data
}
\examples{
# format_watertracker(df)

}
